完成

+++ Collider.h
@@ -6,17 +6,17 @@ public:
 
 	// Getter
 	// 半径を取得
-	float GetRadius() { return radius_; }
+	int GetRadius() { return radius_; }
 
 	// Setter
-	void SetRadius(float radius) { radius_ = radius; }
+	void SetRadius(int radius) { radius_ = radius; }
 
 	virtual void OnCollision() = 0;
 
 	// ワールド座標を取得
-	virtual Vector3 GetWorldPosition() const = 0;
+	virtual Vector3 GetWorldPosition() const =0;
 
 private:
 	// 衝突半径
-	float radius_ = 2;
+	int radius_ = 2;
 };
